+++ GormCore/GormNSPanel.m
@@ -29,7 +29,7 @@
 
 
 // the default style mask we start with.
-static unsigned int defaultStyleMask = NSTitledWindowMask | NSClosableWindowMask
+static NSUInteger defaultStyleMask = NSTitledWindowMask | NSClosableWindowMask
 		  | NSResizableWindowMask | NSMiniaturizableWindowMask;
 
 @implementation GormNSPanel
